wic: Fix image generation for i.MX MXS SoC family
The image needs to use the U-Boot generated, for the uSD card, so a
dependency on that must be added to ensure the file is available prior
'wic' uses it.

Besides that, the fstype needs to be changed so it sets the partition
mark allowing the bootrom to find it.

Series-to: meta-fsl
Series-version: 2
Series-changes: 2
- rework commit log

Series-changes: 3
- fix alignment

Change-Id: I5c7ecdc974aa88c934fed3df2be62a6d7659faf0
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>